OCR폴더에 OCR작업 view에 추가할 함수 일단 넣어놈
